put SPI ROM in cartridge, so per-cart loading/saving is possible, quite a big peformance hit from extra trampolines..
